requirements.txt
fastapi
uvicorn[standard]
psycopg[binary]
sqlalchemy
alembic
python-dotenv
pydantic